refactor resume state management: move initialResumeState to a new file and update import paths
